Phase_2/backend/requirements.txt
fastapi
uvicorn
sqlmodel
pyjwt
python-jose[cryptography]
pydantic-settings
psycopg2-binary
alembic
pytest
httpx
email-validator
